+++ packages/target-tauri/src-tauri/src/autostart.rs
@@ -7,7 +7,8 @@ pub struct AutostartState {
     is_supported: bool,
     // This is not the same as enabled in the desktop settings,
     // this is the actual state not the desktop setting
-    is_registered: bool,
+    // on flatpak this can not be determined
+    is_registered: Option<bool>,
 }
 
 #[cfg(not(desktop))]
@@ -15,11 +16,31 @@ pub struct AutostartState {
 pub(crate) fn get_autostart_state(app: AppHandle) -> Result<AutostartState, String> {
     Ok(AutostartState {
         is_supported: false,
-        is_registered: false,
+        is_registered: Some(false),
     })
 }
 
-#[cfg(desktop)]
+#[cfg(all(desktop, feature = "flatpak"))]
+#[tauri::command]
+pub(crate) async fn get_autostart_state(_app: AppHandle) -> Result<AutostartState, String> {
+    return Ok(AutostartState {
+        is_supported: true,
+        is_registered: None,
+    });
+}
+
+#[cfg(all(desktop, feature = "flatpak"))]
+pub(crate) async fn flatpak_set_auto_start(auto_start: bool) -> Result<(), ashpd::Error> {
+    let response = ashpd::desktop::background::Background::request()
+        .auto_start(auto_start)
+        .command(["deltachat-tauri", "--autostart"])
+        .send()
+        .await;
+    response?.response()?;
+    Ok(())
+}
+
+#[cfg(all(desktop, not(feature = "flatpak")))]
 #[tauri::command]
 pub(crate) fn get_autostart_state(app: AppHandle) -> Result<AutostartState, String> {
     use tauri_plugin_autostart::ManagerExt;
@@ -31,7 +52,7 @@ pub(crate) fn get_autostart_state(app: AppHandle) -> Result<AutostartState, Stri
     let is_supported = true;
     Ok(AutostartState {
         is_supported,
-        is_registered,
+        is_registered: Some(is_registered),
     })
 }